# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/customers": {
      "post": {
        "operationId": "CustomersController_create",
        "summary": "Create a new customer",
        "description": "[Admin] Create a new customer.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The record has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/User"
                    },
                    {
                      "$ref": "#/components/schemas/Customer"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The admin is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "CustomersController_findMany",
        "summary": "Get all the customers by filtering them",
        "description": "[Admin] Get all the filtered customers",
        "parameters": [
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "NIF",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "first-name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "last-name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Role (1: Admin, 2: Store, 3: Customer, 4: Unapproved Store, 5: Unconfirmed, 6: Guest)",
            "schema": {
              "enum": [
                1,
                2,
                3,
                4,
                5,
                6
              ],
              "type": "number"
            }
          },
          {
            "name": "email-confirmed",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "The records have been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/User"
                      },
                      {
                        "$ref": "#/components/schemas/Customer"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "The admin is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          }
        },
        "tags": [
          "Users/Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/customers/pagination": {
      "get": {
        "operationId": "CustomersController_findManyWithPagination",
        "summary": "Get all the customers by filtering them (with pagination)",
        "description": "[Admin] Get all the filtered customers (with pagination).",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Results/page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "NIF",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "first-name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "last-name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Role (1: Admin, 2: Store, 3: Customer, 4: Unapproved Store, 5: Unconfirmed, 6: Guest)",
            "schema": {
              "enum": [
                1,
                2,
                3,
                4,
                5,
                6
              ],
              "type": "number"
            }
          },
          {
            "name": "email-confirmed",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "The records have been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "allOf": [
                              {
                                "$ref": "#/components/schemas/User"
                              },
                              {
                                "$ref": "#/components/schemas/Customer"
                              }
                            ]
                          }
                        },
                        "hasNextPage": {
                          "type": "boolean"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The admin is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          }
        },
        "tags": [
          "Users/Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/customers/{id}": {
      "get": {
        "operationId": "CustomersController_findOne",
        "summary": "Get a customer by ID",
        "description": "[Admin] Get the specified customer by its ID (customer_id).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The record has been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/User"
                    },
                    {
                      "$ref": "#/components/schemas/Customer"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The admin is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          }
        },
        "tags": [
          "Users/Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "CustomersController_update",
        "summary": "Update a customer.",
        "description": "[Admin] Update the specified customer by its ID (customer_id) with the specified data.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The record has been successfully updated."
          },
          "401": {
            "description": "The admin is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CustomersController_remove",
        "summary": "Delete a customer",
        "description": "[Admin] Delete the specified customer by its ID (customer_id).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "mode",
            "required": false,
            "in": "query",
            "description": "Delete mode (soft or hard). Default: soft",
            "schema": {
              "enum": [
                "soft",
                "hard"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The record has been successfully deleted."
          },
          "401": {
            "description": "The admin is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          }
        },
        "tags": [
          "Users/Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/customers/{id}/restore": {
      "patch": {
        "operationId": "CustomersController_restore",
        "summary": "Restore a customer",
        "description": "[Admin] Restore a soft deleted customer specified  by its ID (customer_id).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The record has been successfully restored."
          },
          "401": {
            "description": "The admin is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          }
        },
        "tags": [
          "Users/Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Librería Santa Bárbara API",
    "description": "API documentation for Librería Santa Bárbara",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "The id of the user"
          },
          "user_type_id": {
            "type": "number",
            "example": 1,
            "description": "The id of the user type"
          },
          "role_id": {
            "type": "number",
            "example": 1,
            "description": "The id of the role"
          },
          "email": {
            "type": "string",
            "example": "example@example.com",
            "description": "The email of the user"
          },
          "NIF": {
            "type": "string",
            "example": "12345678A",
            "description": "The NIF of the user"
          },
          "address": {
            "type": "string",
            "example": "Calle Princesa 12B",
            "description": "The address of the user"
          },
          "phone_number": {
            "type": "string",
            "example": "+34661122334",
            "description": "The phone number"
          },
          "email_confirmed": {
            "type": "boolean",
            "example": false,
            "description": "The email confirmation status of the user"
          },
          "hash": {
            "type": "string",
            "example": null,
            "description": "Hash used to verify email at registration"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "example": "2021-01-01",
            "description": "The date when the user was created"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "example": null,
            "description": "The date when the user has been updated"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string",
            "example": null,
            "description": "The date when the user has been (soft) deleted"
          }
        },
        "required": [
          "id",
          "user_type_id",
          "role_id",
          "email",
          "NIF",
          "address",
          "phone_number",
          "email_confirmed",
          "hash",
          "created_at",
          "updated_at",
          "deleted_at"
        ]
      },
...
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "The id of the customer"
          },
          "user_id": {
            "type": "number",
            "example": 1,
            "description": "The id of the user"
          },
          "provider": {
            "type": "string",
            "example": "email",
            "description": "The login provider of the customer"
          },
          "social_id": {
            "type": "string",
            "example": null,
            "description": "The social id associated to a provider of the customer"
          },
          "first_name": {
            "type": "string",
            "example": "John",
            "description": "The first name of the customer"
          },
          "last_name": {
            "type": "string",
            "example": "Doe Doe",
            "description": "The last(s) name of the customer"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "example": null,
            "description": "Date when the customer has been updated"
          }
        },
        "required": [
          "id",
          "user_id",
          "provider",
          "social_id",
          "first_name",
          "last_name",
          "updated_at"
        ]
      },
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "role_id": {
            "type": "number",
            "example": 2,
            "description": "The id of the role that the user will have",
            "enum": [
              1,
              2,
              3,
              4,
              5,
              6
            ]
          },
          "password": {
            "type": "string",
            "example": "test",
            "description": "The password of the user",
            "minLength": 8
          },
          "address": {
            "type": "string",
            "example": "Calle Princesa 12B",
            "description": "The address of the user",
            "maxLength": 300
          },
          "phone_number": {
            "type": "string",
            "example": "+34661122334",
            "description": "The phone number of the user",
            "maxLength": 15
          },
          "email_confirmed": {
            "type": "boolean",
            "example": false,
            "default": false,
            "description": "Whether the email of the user has been confirmed or not by the himself."
          },
          "hash": {
            "type": "string",
            "example": null,
            "default": null,
            "description": "Hash used to verify the email of the user at registration"
          },
          "email": {
            "type": "string",
            "example": "customer@customer.com",
            "description": "The email of the customer. Should be unique among customers.",
            "maxLength": 300
          },
          "NIF": {
            "type": "string",
            "example": "12345678A",
            "description": "The NIF of the customer. Should be alphanumeric and unique among customers.",
            "maxLength": 15
          },
          "provider": {
            "type": "string",
            "example": "email",
            "default": "email",
            "description": "The provider of the customer.",
            "enum": [
              "email"
            ]
          },
          "social_id": {
            "type": "string",
            "example": null,
            "default": null,
            "description": "The social id of the customer associated with the provider."
          },
          "first_name": {
            "type": "string",
            "example": "John",
            "description": "The first name of the customer.",
            "maxLength": 100
          },
          "last_name": {
            "type": "string",
            "example": "Doe Doe",
            "description": "The last name(s) of the customer.",
            "maxLength": 200
          }
        },
        "required": [
          "role_id",
          "email"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "role_id": {
            "type": "number",
            "example": 2,
            "description": "The id of the role that the user will have",
            "enum": [
              1,
              2,
              3,
              4,
              5,
              6
            ]
          },
          "password": {
            "type": "string",
            "example": "test",
            "description": "The password of the user",
            "minLength": 8
          },
          "address": {
            "type": "string",
            "example": "Calle Princesa 12B",
            "description": "The address of the user",
            "maxLength": 300
          },
          "phone_number": {
            "type": "string",
            "example": "+34661122334",
            "description": "The phone number of the user",
            "maxLength": 15
          },
          "email_confirmed": {
            "type": "boolean",
            "example": false,
            "default": false,
            "description": "Whether the email of the user has been confirmed or not by the himself."
          },
          "hash": {
            "type": "string",
            "example": null,
            "default": null,
            "description": "Hash used to verify the email of the user at registration"
          },
          "email": {
            "type": "string",
            "example": "customer@customer.com",
            "description": "The email of the customer. Should be unique among customers.",
            "maxLength": 300
          },
          "NIF": {
            "type": "string",
            "example": "12345678A",
            "description": "The NIF of the customer. Should be alphanumeric and unique among customers.",
            "maxLength": 15
          },
          "provider": {
            "type": "string",
            "example": "email",
            "default": "email",
            "description": "The provider of the customer.",
            "enum": [
              "email"
            ]
          },
          "social_id": {
            "type": "string",
            "example": null,
            "default": null,
            "description": "The social id of the customer associated with the provider."
          },
          "first_name": {
            "type": "string",
            "example": "John",
            "description": "The first name of the customer.",
            "maxLength": 100
          },
          "last_name": {
            "type": "string",
            "example": "Doe Doe",
            "description": "The last name(s) of the customer.",
            "maxLength": 200
          }
        }
      },
...
    }
  }
}